+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## _(in-progress)_
+* Hardening: Escape filtered and translated data before output
+
 ## 1.6.2 _(2024-08-21)_
 * Fix: Fix formatting of code in `readme.txt`
 * Change: Note compatibility through WP 6.6+